feat: add menubar
