+++ src/routes/admin/clusters/[cluster_uuid]/select/+page.ts
